feat: RSC+SSR react build (#53)
## It works!
<img width="516" alt="Screenshot 2025-01-21 at 17 19 12" src="https://github.com/user-attachments/assets/f60ec7c7-2eb9-49dd-b23d-6e1ed17fe6c1" />

^ PDF page that previously failed with `useCallback()` error in `react-to-print` 

## Videos with context
1. [Problem + Solution Overview](https://www.loom.com/share/bff62948fcae4c4fa128638ab1039fd8?sid=832f0f6f-1e9a-4b2a-a557-2f3223f1e5bb)
2. [Why not proxy between RSC/SSR react dynamically?](https://www.loom.com/share/e028884d6eae4e9c957b9a73cbeff2e9?sid=d9e1318f-51df-4f57-b6d3-0b1e991d5bf8)
3. [Why not determine which react to use when bundling?](https://www.loom.com/share/ebc61178938c45a380dc22f2b99b04b4?sid=57f03e98-d3be-42c0-9c48-77e39ef2da26)
4. [Why not route SSR rendering requests to a different worker?](https://www.loom.com/share/4d2fb53835fd47cbb524117ca52cb1f0?sid=8db936d2-3b83-4013-ac96-c5e92d4339d5)

## Problem

RSC React and SSR React need to coexist in the same Cloudflare Worker. The challenge arises when dependencies import React - unlike our own code (which previous solutions focused on), in this context we do not know whether to use the SSR or RSC runtime.

## Solution

This PR introduces a **custom React build** that is mostly just a standard React build, but with both server and client internals exposed.

RSC React is essentially a **subset** of SSR React—their core code is largely the same. The only difference is how each exposes internal objects:

* **SSR React** exposes _client internals_

* **RSC React** exposes _server internals_

React DOM uses these internals to determine whether it’s in SSR or RSC mode, meaning the runtime itself doesn’t need to be switched—React simply checks the available internals.

By creating a custom build that exports both sets of internals, we eliminate the need to juggle runtimes dynamically. React functions as a single runtime that works seamlessly across both SSR and RSC contexts.

## Why not X?

1. **Proxying between RSC/SSR at runtime**
* Difficult to manage context reliably
* React’s internal event loop makes it difficult to set which context we should be in (even with things like [[AsyncLocalStorage](https://nodejs.org/api/async_context.html#troubleshooting-context-loss)](https://developers.cloudflare.com/workers/runtime-apis/nodejs/asynclocalstorage/))
* Increased complexity with minimal benefits

2. **Determining the runtime at build time**
* Works for project code but fails for dependencies used across both SSR and RSC
* Requires separate bundles, increasing complexity and size

3. **Routing SSR requests to a separate worker**
* Still requires changing the react runtime dynamically, leading to the similar challenges as (1)
* Adds operational overhead without solving the core issue

For more, check out the videos!