fix: Avoid while loop for `pnpm dev` (#154)
## Context
Early implementations of the sdk relied on us having `vite dev` in a ` while true` bash loop, as a way to ensure changes to configuration restarted the dev server.

We don't need this anymore, but the starters still have this remnant around. This PR removes it so that `pnpm dev` is basically `pnpm vite dev`