chore: Check starters on CI (#192)
## CI checks
To be more sure we're giving people working code to base their applications on, this PR adds a github workerflow that run the following on CI for each starter:
* ts typechecks
* pnpm build
* pnpm dev and check that cURL-ing it comes back with a 200
* pnpm preview and check that cURL-ing it comes back with a 200

## New `package.json` scripts in starters
* `pnpm check`: runs tsc (users can change it to do other checks too if desired)
* `pnpm preview`: builds then run `vite preview`

## No more port setting
We no longer set a port (2332) for the dev server - this was a remnant of where we started when building the sdk that we don't need anymore. This means `vite dev` will now use its own default port (5173), and that users can do `pnpm dev --port <PORT>` to override this.

## `dev:init` is now db agnostic
As long as a project has `migrate:dev` and `seed` scripts  in their `package.json`, we'll run them on postinstall (previously we only ran them if we found a prisma schema). This means the drizzle starter also only needs `pnpm i && pnpm dev` to get a running server