chore: Rename `layout()` -> `document()` (#193)
We currently have a `layout()` API method, which lets the user define the "shell" of their application's html: the stuff that goes in their `<head>` and `<body>`, and where the actual page fits into the body.

The document content itself is server-side rendered and served to the user's browser. Within this html, the user chooses the mount point where their actual page content will be situated. This page content is the "dynamic" stuff which updates using React RSC.

Behaviour wise, this is different to `layout()` as a concept as it is defined in React Router.

To avoid confusing users familiar with React Router, we're renaming `layout()` to `document()`.