fix: Use pnpm dev for dev:init when no .wrangler dir is found (#210)
## Context
We want the user experience for getting started to be as simple as:
```shell
pnpm i
pnpm dev
```

To do this, we currently have a `postinstall` hook in the starters that runs `pnpm dev:init`, which in turn will run `pnpm migrate:dev`, then `pnpm seed`.

## Problem
The first problem is that getting this to work properly currently involved some workarounds that I'm not too happy about: #208, #209.

The second problem is that our repo now takes long to `pnpm install` each time, since every time we are needing to run migrate commands and seed commands for each starter.

The last problem (arguably worst) is that every time users `pnpm install`, we'd be re-running their seed script. They may not want to re-run their seed script - e.g. it would probably blow away all their dev data they were working on - unless they specifically made their seed script not do this.

## Solution
Thinking about it some more, `postinstall` isn't really the place we want `pnpm dev:init` to happen. We want to happen on _fresh installs_ - not on re-installs.

Fortunately, there's a reliable way to determine if it is a fresh dev install - whether there is a `.wrangler` dir in the project. If there isn't, for one it means migrations have not yet been run (=> seed data not there yet either).

This PR changes things such that when the user runs `pnpm dev`, if there is no `.wrangler`, we `pnpm dev:init` first.

We also change the package.json scripts of the starters so that they can still run `pnpm check` now that the postinstall doesn't `dev:init` (we just needed to make sure we run prisma generate before typechecking)