fix: HMR for realtime (#234)
## Problem
We send an "empty" action request in order when we detect an HMR update is needed.

However, for realtime, we need to be able to forward the action request from the DO (which manages the websockets) to the worker. To this, we need a JSON serializable value for the action id and args.

## Solution
Send an internal action id (`__rsc_hot_update`) instead of an empty message, and special case this to do nothing when received server side (we only care about the new rsc result, not the action).