feat: Avoid manual deploy setup + env hoop jumping (#304)
## Context
A significant friction point in our DX is around env vars and deployment setup. Ideally it should be just as simple to do first deploy without any input, as it is do the same in dev. This PR solves this as follows.

## `npm run release` now has no manual steps
Before deploying, we now run a script that works as follows:
* if it finds `env.DB` usage in `src/`, it creates a d1 db (unless one has already been created)
* if it finds `WEBAUTHN` usage in `src/`, it creates the necessary auth env vars and sets them with `wrangler secret`

## Turnstile is now opt-in
The problem with turnstile is that there is unfortunately a manual step in CF UI. For this reason we leave Turnstile out, and recommend in the docs how to opt into it, and why you would want to

## No more need for `.env` in dev for auth
We now have defaults for the auth vars in dev, so you don't need an `.env` file unless you want to define your own APP-specific env vars.

## Links
* Discussion showing the current pain points: https://discord.com/channels/679514959968993311/1307974274145062912/1360728911071150202

Fixes #302, #284, #309