fix: Separate webauthn vars for auth secret key vars in deployment script (#329)
In our deployment script, we check if the codebase has webauthn references, and only then set the relevant vars. At the moment this also includes `AUTH_SECRET_KEY` - but this is problematic since an application might have no webauthn but still reference `AUTH_SECRET_KEY`.

This PR fixes this by checking and handling for AUTH_SECRET_KEY separately in the deployment script.