fix: Check for sdk/auth import in deploy script (#330)
In #329, we're looking for an `AUTH_SECRET_KEY` when deciding whether to set it during deployment. Except we no longer require this var to be specified in user app (we look it up in the sdk itself). What we _should_ be looking for, is imports to `@redwoodjs/sdk/auth`. That's what we change in this PR.