chore: Move react deps to sdk (#331)
## Context
While RSC is still stabilizing, it is helpful for us to have more control over what versions of react are being used. e.g. if regressions are introduced in new versions, based on how we use RSC.

For this reason, we are moving react deps into the sdk.

## Implementation
### Problem
We need to put the react deps in vite's `optimizeDeps` to convert them from cjs to esm. Problem is, vite can't find the deps once we've moved them to the sdk. It is not enough to use the `dep > subdep` feature of `optimizeDeps`:
* it does not appear to work for cases where there are import conditions (e.g. `react/jsx-runtime`)
* even if there is a way to get it working, there are some plugins (e.g. vite react plugin) that simply include these deps in `optimizeDeps`, and we aren't able to override this

### Solution
`optimizeDeps` looks at aliases when determining how to resolve the deps in `optimizeDeps.include`: https://github.com/vitejs/vite/blob/fdb36e076969c763d4249f6db890f8bf26e9f5d1/packages/vite/src/node/idResolver.ts#L62

This PR leverages this - it adds a vite plugin that resolves the exact paths to each of the react deps, and then puts these in aliases.

This also makes it possible to make two further improvements
* We can now control dev vs production versions of react deps (Fixes #292)
* We no longer need a custom build for react dom server edge (we now only have a custom build for react itself)

## ⚠️ Migration recommendation
While not strictly a breaking change, now that the sdk has react packages as its own dependencies, it is possible for strange things to happen if there are also versions of the react deps in your own project as well.

This is easily solved by simply removing these dependencies from your `package.json` and re running npm/pnpm/yarn install after upgrading to the latest sdk version.

```diff
-    "react": "canary",
-    "react-dom": "canary",
-    "react-server-dom-webpack": "canary"
```