fix: React dep resolution should respect import conditions correctly (#339)
## Context
In order to have control over the react versions we use (while RSC is stabilizing), we maintain the react versions in the sdk package itself rather than users having react packages as deps in their own apps. To accomplish this involves needing to tell vite that when doing `optimizeDeps` (since we need to de-cjs-ify react packages), that it should look for these deps as deps of the sdk instead. More on this in the initial PR where this was introduced - #331

## Problem
In the initial PR where this was introduced (#331), we used some complicated logic of our own for determining which react package to use. We needed this to have control over when to use the packages that correspond to `react-server` import conditions vs not. There were bugs in this implementation that (for one) were causing us to incorrectly resolve `react-dom` imports in the `worker` vite environment to the non `react-server` package.

## Solution
What we are needing is an import-condition-aware resolver. There are already ones out there. This PR replaces our (my) broken custom resolution logic with the resolver used under the hood by webpack: https://www.npmjs.com/package/enhanced-resolve