fix: Better support for redirect responses (#342)
## Changes
### Fix types: `MaybePromise<Response | Element>`
Previously if a component returned either a `Response` or a react element, ts would tell you the function should return one or the other but not both. Now we allow `MaybePromise<Response | Element>`

### Fix runtime for route handlers
Previously, if we detected that a function returned react elements, we did not allow for the possibility of it also potentially returning a response. We now handle both react elements or responses being returned.

### Support throwing responses
We currently do not have a way to bubble up responses nested in the component tree. This PR allows ones to do this:

```tsx
export const Child = () => {
  if (someCondition) {
    return new Response(null, {
      status: 302,
      headers: {
        Location: "/here",
      },
    });
  }

  return <div>hi</div>
};

export const Page = () => (
  <>
    <Child />
  <>
);
```