docs: Environment Variables - Runtime Types (#383)
When I add new envars, I often forget that I have to run `pnpm wrangler types`.

While the envar is loaded fine, the types are unknown red squiggles.

Added this to the envar docs to explain how to and why.

Question: Should the starters also have a package script to do this?