fix: More robust check for prisma at build time (#403)
## Context
We check whether the project is using prisma in order to apply build configuration specific to prisma - largely around getting prisma query engine wasm to work with deploys to cloudflare when using `@cloudflare/vite-plugin`.

## Problem
The way we (I) checked whether the project is using prisma was pretty awful: `pnpm prisma --version`
* If you're not using pnpm or corepack, we'll incorrectly take it that you're not using prisma
* If for some reason you have a prisma global install but aren't using prisma for you're project, we'll incorrectly take it that you are using prisma
* There's probably further ways this check can break

## Solution
Check whether the project has `@prisma/client` as an installed dependency by trying to resolve it (using https://www.npmjs.com/package/enhanced-resolve). If it can be resolved, we take it that you're using prisma. If the resolve returns a falsy value or throws, we can it that you aren't using prisma.


Fixes #396