fix: Also look for *.tsx in directive transformation during optimizeDeps (#511)
## Problem
In order to support `"use client"` in `node_modules`, we also have an esbuild plugin for vite's optimizeDeps step that does the same `"use client"` directive transformation as we do for non-deps. In order for this to work, vite also needs to run the app code through `optimizeDeps` in order to discovery the correct deps.

For example, we should only be discovery client deps for ssr and client environments and not for worker environment, so we need to apply the relevant transformations for each environment so that we don't end up processing client deps in worker environment (since in use client case for worker environment, we transform the module to basically just be a thin rpc layer RSC provides - i.e. no dep imports).

We were doing all of this fine. It is just we weren't also looking for .*.ts and *.tsx files when doing this transformation during optimizeDeps.

## Solution
Add *.ts and *.tsx to the filter we use in our optimizeDeps plugin for directive transformation.