feat: Avoid ts-morph parse for non-"use server" cases (#520)
In order to support RSC for the full module graph (including `node_modules`), we need to run the entire module graph through the `"use client"` and `"use server"` transformations. That said, we can still skip transformations as soon as we can tell that the given module has not "use server" or "use client" directive.

At the moment we exit early this way for "use client" - we first do a simple string check that the code starts with "use client" before parsing the code with ts-morph.

We currently do not do the same for "use server", which means every module in the graph (whether it is app code or code in node_modules) is first parsed with ts-morph.

This PR does the same as "use client" in the "use server" transform: first do a simple string check to test whether the code starts with "use server", if it does then parse with ts-morph, otherwise skip.