fix: Avoid misleading wrapper error message (#584)
I added wrapper error info in #556 so that we include component stack. Part of that wrapper included a `Error Rendering RSC to HTML stream.` prefix to the underlying error message.

That prefixing ended up misleading people into thinking the actual error was `Error Rendering RSC to HTML stream` (#564). This PR removes that prefix to avoid this, but keeps the component stack info.