fix: SSR on explicit reloads after HMR (#586)
## Context
We have a concept of virtual SSR modules as part of the worker environment's module graph. Each corresponds to a module in the SSR environment's module graph.

This allows us to do processing for two different environments - most importantly, `worker` environment can use `react-server` import condition and RSC react runtime, `ssr` environment can use classic react runtime and _no_ `react-server` import condition.

In order to accomplish this, we have a vite `load()` plugin hook, which loads the virtual SSR modules in the worker env by calling `fetchModule` for the corresponding module in the SSR env.

## Problem 1
We need to invalidate the virtual SSR module when the actual file changes, so that vite can reload it (as well as dependent modules in the graph) correctly with its changes accounted for.

## Problem 2: The underlying one
Invalidating it should be enough - if we actually had the correct module graph.

Turns out we didn't.

We had this - notice that each virtual module is a leaf node:

```
- /Users/justin/rw/blotter/rwsdk-hydration-example/src/worker.tsx
  ...
- virtual:rwsdk:ssr:/node_modules/.vite/deps_ssr/chunk-5SVCMTR3.js?v=6970a862 [virtual]
- virtual:rwsdk:ssr:/node_modules/.vite/deps_ssr/chunk-6Y5HNGJW.js?v=6970a862 [virtual]
- virtual:rwsdk:ssr:/node_modules/.vite/deps_ssr/chunk-TBJU3W5Y.js?v=6970a862 [virtual]
- virtual:rwsdk:ssr:/node_modules/.vite/deps_ssr/chunk-5UXEC3AS.js?v=6970a862 [virtual]
- virtual:rwsdk:ssr:/node_modules/.vite/deps_ssr/chunk-BDXQCHU4.js?v=6970a862 [virtual]
- virtual:rwsdk:ssr:/node_modules/.vite/deps_ssr/chunk-PTYHVFOF.js?v=6970a862 [virtual]
- virtual:rwsdk:ssr:/@id/virtual:use-client-lookup.js [virtual]
- virtual:rwsdk:ssr:/src/app/pages/components/client-component.tsx [virtual]
- virtual:rwsdk:ssr:/node_modules/.vite/deps_ssr/react_jsx-dev-runtime.js?v=7d86d101 [virtual]
- virtual:rwsdk:ssr:/src/app/pages/components/foo.ts [virtual]
```

What we want is this:

```
- /Users/justin/rw/blotter/rwsdk-hydration-example/src/worker.tsx
...
    - virtual:rwsdk:ssr:rwsdk/__ssr_bridge [virtual]
      - virtual:rwsdk:ssr:/node_modules/.vite/deps_ssr/chunk-PTYHVFOF.js?v=fb8dc2f5 [virtual]
      - virtual:rwsdk:ssr:/@id/virtual:use-client-lookup.js [virtual]
        - virtual:rwsdk:ssr:/src/app/pages/components/client-component.tsx [virtual]
          - virtual:rwsdk:ssr:/src/app/pages/components/foo.ts [virtual]
        - virtual:rwsdk:ssr:/src/app/pages/user/Login.tsx [virtual]
...
```

Why is it a problem that graph is incorrect? Vite correctly identifies the invalidate as a dead end (since there are no importers of the module), and does a full reload.

Why didn't we have the correct module graph?

The `fetchModule` code we do in the `load()` for the virtual SSR modules had imports that look like this, as a result of vite transforming them:

```ts
__vite_ssr_import__("...")
```

These are fine for runtime purposes, and at this point vite has already tracked the imports it needs for the relevant environment (`ssr`) before it did this transforming.

Thing is, we're now returning this code in a `load()` that is processing for a _different environment_ - `worker`. Vite needs to see `import`s instead so that it can track them correctly to build the module graph.

## Solution
* To build module graph correctly: use ast-grep to collect all `__vite_ssr_import__`s and `__vite_ssr_dynamic_import__`s
* Wrap the code such that whenever a `_vite_ssr_import` or `__vite_ssr_dynamic_import__` happens, we also `import()` it as a side effect - importantly, the literal itself. The resulting code looks something like
```ts
await (async function(__vite_ssr_import__, __vite_ssr_dynamic_import__) {/* code returned by fetchModule() goes here */ })(
  (id, ...args) => ssrImport(id, false, ...args),
  (id, ...args) => ssrImport(id, true, ...args)
);

function ssrImport(id, isDynamic, ...args) {
  switch (id) {
    case "/src/app/pages/components/client-component.ts": import("virtual:rwsdk:ssr:/src/app/pages/components/client-component.ts"),
    case "/src/app/pages/components/foo.ts": import("virtual:rwsdk:ssr:/src/app/pages/components/foo.ts"),
    // ...
  }

  const virtualId = '/@id/virtual:rwsdk:ssr:' + id;
  return isDynamic ? __vite_ssr_dynamic_import__(virtualId, ...args) : __vite_ssr_import__(virtualId, ...args);
}
```
* Then invalidate the virtual SSR module when the corresponding actual file changes

## Links
* [Support thread](https://discord.com/channels/679514959968993311/1392212607322034187)