fix: Adding "use client" after dev server running (#589)
## Problem
Currently, if a module starts off as without "use client", then gets "use client" added to it once the dev server is running, you'll get an error:
```
(ssr) No module found for '<id>' in module lookup for "use client" directive
```

The reason for this, is that we need to invalidate the client lookup (we use this module for finding the relevant client components when doing client or server side rendering), once a file gets a "use client" directive.

## Solution
Check whether files go from having to not having "use client" or "use server" directives, update the set of files we track accordingly, then invalidate the client (or server) lookup module so that it can be updated.

## Demo of fix
https://www.loom.com/share/ba1a0f1c1e5949eba81b5c86a38aebdd