fix: HMR for adding/removing "use client" (#637)
We maintain lookups of client and server modules to use for client and ssr to be able to find and import these modules dynamically when they need to be used.

We make sure to add/remove from these lookups as files change (e.g. `"use client"` getting added to a module that did not have it). Problem is, when we did this adding/removing, we were not normalizing the paths - the lookups use vite style paths (`/src/app/Button.tsx`), but we were setting them as absolute paths (`/path/to/project/src/app/Button.tsx`)