fix: Request cascade for client components (#649)
### Problem

Our server knows everything, but the browser starts with nothing.

As part of the RSC render, our server dynamically discovers the exact set of "use client" components and stylesheets needed for any given page. It has perfect information before the request is even finished. The problem is that this knowledge is trapped on the server.

The browser, on the other hand, receives an HTML document and has to work backward. It's forced into a request waterfall:
1.  Fetch the document.
2.  Find, fetch, and execute the main entry script.
3.  *Only then* can it begin to render and learn about the client components it needs, triggering a second wave of network requests.

Each step in this chain blocks the next. This sequential loading pattern directly harms performance, delaying when a user can interact with the page (Time-to-Interactive) and (when there's no SSR in the mix) when the main content becomes visible (Largest Contentful Paint).

**Before:**
<img width="1258" height="669" alt="before" src="https://github.com/user-attachments/assets/8e8d48d6-fd3e-4185-bd57-25e87e65e785" />

### Solution

The key to solving this lies in a mechanism we already have. To prevent a flash of unstyled content, we already determine the exact set of scripts needed for a page to find all their relevant stylesheet dependencies. This PR now uses that same list of scripts for a second purpose: to generate preload hints.

We now inject `<link rel="modulepreload">` tags for the JavaScript modules directly into the server's response stream. This is made possible by React 19's automatic hoisting of `<link>` tags into the document `<head>`, which ensures the browser sees these hints immediately.

<img width="441" height="45" alt="preload" src="https://github.com/user-attachments/assets/0f3124b6-1bab-4f54-ad2c-9cfe8d9255fc" />


The result is a flattened request waterfall, as the browser can now fetch all necessary assets in parallel with the main client entry point.

**After:**
<img width="1240" height="575" alt="after" src="https://github.com/user-attachments/assets/208e3bec-c0e4-4cb1-8155-34497c92c0ad" />