feat: Worker run and redwood plugin improvements (#661)
## Problem
Currently, one is able to specify a custom config path for their wrangler config in our vite plugin, but not with `worker-run` scripts. We need to be able to specifying the wrangler config in both cases.

## Solution
Change both the vite pugin and `worker-run` to make use of a `RWSDK_WRANGLER_CONFIG` env var to specify the wrangler config path.

Went fr env var instead of a cli arg, so that it can be propagated more easily to the different parts that need to know the worker config path.

```bash
RWSDK_WRANGLER_CONFIG=... pnpm dev:init
RWSDK_WRANGLER_CONFIG=... pnpm dev
```

Solves #636