fix: Store module env correctly in directive scanner
A previous refactoring in #675 to clean up the directive scanner was overly aggressive. It removed logic from the `esbuild` plugin's `onLoad` hook that was responsible for determining and caching a module's environment. By correctly determining and caching a module's own environment when it is loaded, we ensure that the resolution context is propagated correctly throughout the entire dependency graph.