feat: add smoke test matrix for package managers and OSes
Implements a testing matrix in the smoke test CI workflow.

The smoke tests will now run against a combination of:
- Starters: minimal, standard
- Operating Systems: ubuntu-latest, windows-latest
- Package Managers: pnpm, npm, yarn, yarn-classic

This change introduces a  flag to the smoke test
runner script to facilitate this matrix. The test environment setup
is now capable of installing dependencies using the specified package
manager.